function given as a parameter on each element of an array.

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,5 @@
-#include <stdio.h>
 #include "function_pointers.h"
+#include <stdio.h>
 
 /**
  * array_iterator - run a function given as a parameter on each array element
@@ -10,7 +10,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int  i;
 
 	if (array != NULL && action != NULL)
 	{
